+++ src/main/java/com/oceanprotocol/squid/api/AssetsAPI.java
@@ -17,6 +17,7 @@ import com.oceanprotocol.squid.models.asset.OrderResult;
 import com.oceanprotocol.squid.models.service.ProviderConfig;
 import io.reactivex.Flowable;
 
+import java.io.InputStream;
 import java.util.Map;
 
 /**
@@ -90,6 +91,31 @@ public interface AssetsAPI {
     public SearchResult query(Map<String, Object> params) throws DDOException;
 
     /**
+     *  Downloads a single file of an Asset previously ordered through a Service Agreement
+     * @param serviceAgreementId the service agreement id of the asset
+     * @param did the did
+     * @param serviceDefinitionId the service definition id
+     * @param index of the file inside the files definition in metadata
+     * @param basePath  the path where the asset will be downloaded
+     * @param threshold secret store threshold to decrypt the urls of the asset
+     * @return a flag that indicates if the consume flow was executed correctly
+     * @throws ConsumeServiceException ConsumeServiceException
+     */
+    public Boolean consume(String serviceAgreementId, DID did, String serviceDefinitionId, Integer index, String basePath, int threshold) throws ConsumeServiceException;
+
+    /**
+     *  Downloads a single file of an Asset previously ordered through a Service Agreement
+     * @param serviceAgreementId the service agreement id of the asset
+     * @param did the did
+     * @param serviceDefinitionId the service definition id
+     * @param index of the file inside the files definition in metadata
+     * @param basePath the path where the asset will be downloaded
+     * @return a flag that indicates if the consume flow was executed correctly
+     * @throws ConsumeServiceException ConsumeServiceException
+     */
+    public Boolean consume(String serviceAgreementId, DID did, String serviceDefinitionId, Integer index, String basePath) throws ConsumeServiceException;
+
+    /**
      * Downloads an Asset previously ordered through a Service Agreement
      * @param serviceAgreementId the service agreement id of the asset
      * @param did the did
@@ -111,6 +137,10 @@ public interface AssetsAPI {
      * @throws ConsumeServiceException ConsumeServiceException
      */
     public Boolean consume(String serviceAgreementId, DID did, String serviceDefinitionId, String basePath) throws ConsumeServiceException;
+
+
+    public InputStream consumeBinary(String serviceAgreementId, DID did, String serviceDefinitionId, Integer index) throws ConsumeServiceException;
+    public InputStream consumeBinary(String serviceAgreementId, DID did, String serviceDefinitionId, Integer index, int threshold) throws ConsumeServiceException;
 
     /**
      * Purchases an Asset represented by a DID. It implies to initialize a Service Agreement between publisher and consumer